# Optimized std.datetime.systime.SysTime.fromSimpleString