Much less to do now that `find` no longer auto decodes thanks to https://github.com/dlang/phobos/pull/6132.

```
$ dmd/generated/osx/release/64/dmd -O -inline -release test.d && ./test
2010-Jul-04 07:06:12+08:00
old		8 secs, 791 ms, and 976 μs
new		4 secs, 685 ms, 384 μs, and 9 hnsecs
```

<details>

```d
import std.stdio;
import std.exception;
import std.conv;
import std.ascii;
import std.range;
import std.traits;
import std.typecons;
import std.string;
import std.datetime.systime;
import std.datetime.date;
import std.datetime.timezone;
import std.datetime.stopwatch;
import core.time;
import std.random;

enum testCount = 5_000_000;

__gshared a = "2010-Jul-04 07:06:12+08:00";

SysTime fromSimpleString1(S)(in S simpleString, immutable TimeZone tz = null) @safe
        if (isSomeString!(S))
{
    import std.algorithm.searching : countUntil, find;
    import std.conv : to;
    import std.string : strip;

    auto dstr = to!dstring(strip(simpleString));

    auto spaceIndex = dstr.countUntil(' ');
    enforce(spaceIndex != -1, new DateTimeException(format("Invalid Simple String: %s", simpleString)));

    auto found = dstr[spaceIndex + 1 .. $].find('.', 'Z', '+', '-');
    auto dateTimeStr = dstr[0 .. $ - found[0].length];

    dstring fracSecStr;
    dstring zoneStr;

    if (found[1] != 0)
    {
        if (found[1] == 1)
        {
            auto foundTZ = found[0].find('Z', '+', '-');

            if (foundTZ[1] != 0)
            {
                fracSecStr = found[0][0 .. $ - foundTZ[0].length];
                zoneStr = foundTZ[0];
            }
            else
                fracSecStr = found[0];
        }
        else
            zoneStr = found[0];
    }

    try
    {
        auto dateTime = DateTime.fromSimpleString(dateTimeStr);
        auto fracSec = fracSecsFromISOString(fracSecStr);
        Rebindable!(immutable TimeZone) parsedZone;

        if (zoneStr.empty)
            parsedZone = LocalTime();
        else if (zoneStr == "Z")
            parsedZone = UTC();
        else
            parsedZone = fromISOExtStringS(zoneStr);

        auto retval = SysTime(dateTime, fracSec, parsedZone);

        if (tz !is null)
            retval.timezone = tz;

        return retval;
    }
    catch (DateTimeException dte)
        throw new DateTimeException(format("Invalid Simple String: %s", simpleString));
}

SysTime fromSimpleString2(S)(in S simpleString, immutable TimeZone tz = null) @safe
        if (isSomeString!(S))
{
    import std.algorithm.searching : find;
    import std.conv : to;
    import std.string : strip, indexOf;

    auto str = strip(simpleString);

    auto spaceIndex = str.indexOf(' ');
    enforce(spaceIndex != -1, new DateTimeException(format("Invalid Simple String: %s", simpleString)));

    auto found = str[spaceIndex + 1 .. $].byCodeUnit.find('.', 'Z', '+', '-');
    auto dateTimeStr = str[0 .. $ - found[0].length];

    typeof(str) fracSecStr;
    typeof(str) zoneStr;

    if (found[1] != 0)
    {
        if (found[1] == 1)
        {
            auto foundTZ = found[0].find('Z', '+', '-');

            if (foundTZ[1] != 0)
            {
                static if (isNarrowString!S)
                {
                    fracSecStr = found[0][0 .. $ - foundTZ[0].length].source;
                    zoneStr = foundTZ[0].source;
                }
                else
                {
                    fracSecStr = found[0][0 .. $ - foundTZ[0].length];
                    zoneStr = foundTZ[0];
                }
            }
            else
            {
                static if (isNarrowString!S)
                    fracSecStr = found[0].source;
                else
                    fracSecStr = found[0];
            }
        }
        else
        {
            static if (isNarrowString!S)
                zoneStr = found[0].source;
            else
                zoneStr = found[0];
        }
    }

    try
    {
        auto dateTime = DateTime.fromSimpleString(dateTimeStr);
        auto fracSec = fracSecsFromISOString(fracSecStr);
        Rebindable!(immutable TimeZone) parsedZone;

        if (zoneStr.empty)
            parsedZone = LocalTime();
        else if (zoneStr == "Z")
            parsedZone = UTC();
        else
            parsedZone = fromISOExtStringS(zoneStr);

        auto retval = SysTime(dateTime, fracSec, parsedZone);

        if (tz !is null)
            retval.timezone = tz;

        return retval;
    }
    catch (DateTimeException dte)
        throw new DateTimeException(format("Invalid Simple String: %s", simpleString));
}

auto byCodeUnit(R)(R r)
if (isAutodecodableString!R ||
    isInputRange!R && isSomeChar!(ElementEncodingType!R) ||
    (is(R : const dchar[]) && !isStaticArray!R))
{
    static if (isNarrowString!R ||
               // This would be cleaner if we had a way to check whether a type
               // was a range without any implicit conversions.
               (isAutodecodableString!R && !__traits(hasMember, R, "empty") &&
                !__traits(hasMember, R, "front") && !__traits(hasMember, R, "popFront")))
    {
        static struct ByCodeUnitImpl
        {
        @safe pure nothrow @nogc:

            @property bool empty() const     { return source.length == 0; }
            @property auto ref front() inout { return source[0]; }
            void popFront()                  { source = source[1 .. $]; }

            @property auto save() { return ByCodeUnitImpl(source.save); }

            @property auto ref back() inout { return source[$ - 1]; }
            void popBack()                  { source = source[0 .. $-1]; }

            auto ref opIndex(size_t index) inout     { return source[index]; }
            auto opSlice(size_t lower, size_t upper) { return ByCodeUnitImpl(source[lower .. upper]); }

            @property size_t length() const { return source.length; }
            alias opDollar = length;

            StringTypeOf!R source;
        }

        static assert(isRandomAccessRange!ByCodeUnitImpl);

        return ByCodeUnitImpl(r);
    }
    else static if (is(R : const dchar[]) && !__traits(hasMember, R, "empty") &&
                    !__traits(hasMember, R, "front") && !__traits(hasMember, R, "popFront"))
    {
        return cast(StringTypeOf!R) r;
    }
    else
    {
        // byCodeUnit for ranges and dchar[] is a no-op
        return r;
    }
}

Duration fracSecsFromISOString(S)(in S isoString) @trusted pure
if (isSomeString!S)
{
    import std.algorithm.searching : all;
    import std.ascii : isDigit;
    import std.conv : to;
    import std.string : representation;

    if (isoString.empty)
        return Duration.zero;

    auto str = isoString.representation;

    enforce(str[0] == '.', new DateTimeException("Invalid ISO String"));
    str.popFront();

    enforce(!str.empty && all!isDigit(str), new DateTimeException("Invalid ISO String"));

    dchar[7] fullISOString = void;
    foreach (i, ref dchar c; fullISOString)
    {
        if (i < str.length)
            c = str[i];
        else
            c = '0';
    }

    return hnsecs(to!int(fullISOString[]));
}

immutable(SimpleTimeZone) fromISOExtStringS(S)(S isoExtString) @safe pure
        if (isSomeString!S)
{
    import std.algorithm.searching : startsWith;
    import std.conv : ConvException, to;
    import std.datetime.date : DateTimeException;
    import std.exception : enforce;
    import std.format : format;
    import std.string : indexOf;

    auto whichSign = isoExtString.startsWith('-', '+');
    enforce!DateTimeException(whichSign > 0, format("Invalid ISO String: %s", isoExtString));
    auto sign = whichSign == 1 ? -1 : 1;

    isoExtString = isoExtString[1 .. $];
    enforce!DateTimeException(!isoExtString.empty, format("Invalid ISO String: %s", isoExtString));

    immutable colon = isoExtString.indexOf(':');
    S hoursStr;
    S minutesStr;
    int hours, minutes;

    if (colon != -1)
    {
        hoursStr = isoExtString[0 .. colon];
        minutesStr = isoExtString[colon + 1 .. $];
        enforce!DateTimeException(minutesStr.length == 2, format("Invalid ISO String: %s", isoExtString));
    }
    else
    {
        hoursStr = isoExtString;
    }

    enforce!DateTimeException(hoursStr.length == 2, format("Invalid ISO String: %s", isoExtString));

    try
    {
        // cast to int from uint is used because it checks for
        // non digits without extra loops
        hours = cast(int) to!uint(hoursStr);
        minutes = cast(int) (minutesStr.empty ? 0 : to!uint(minutesStr));
    }
    catch (ConvException)
    {
        throw new DateTimeException(format("Invalid ISO String: %s", isoExtString));
    }

    enforce!DateTimeException(hours < 24 && minutes < 60, format("Invalid ISO String: %s", isoExtString));

    return new immutable SimpleTimeZone(sign * (dur!"hours"(hours) + dur!"minutes"(minutes)));
}

void main()
{
    SysTime res;

    auto result = to!Duration(benchmark!(() => res = fromSimpleString1(a))(testCount)[0]);
    auto result2 = to!Duration(benchmark!(() => res = fromSimpleString2(a))(testCount)[0]);

    writeln(res); // force execution

    writeln("old", "\t\t", result);
    writeln("new", "\t\t", result2);
}
```

</details>