@wilzbach 2018-02-14
(DScanner failure is unrelated. Fix: https://github.com/dlang/phobos/pull/6171)